% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel2num.R
\name{cel2num}
\alias{cel2num}
\title{Convert Excel cell index to numeric column and row indexes}
\usage{
cel2num(celIndex, rrw = 0, rcn = 0)
}
\arguments{
\item{celIndex}{with maxmum 2-letter column and a number, e.g. Ad23.}

\item{rrw}{a (+/-) integer relative to celIndex row, defaults to 0.}

\item{rcn}{a (+/-) integer relative to celIndex column, defaults to 0.}
}
\value{
The number of row with cel2num(...)$r or cel2num(...)[1], and the number of column with cel2num(...)#c or cel2num(...)[2].
}
\description{
This function converts an Execl cell index (with max 2 column letters) to row and column numbers.
}
\examples{
cel2num("mD263")$r or cel2num("mD263")[1] for row number
cel2num("mD263")$c or cel2num("mD263")[2] for column number
}
\keyword{Excel}
